tools/backlink-bot/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
browser-use
playwright
google-generativeai
langchain-google-genai
aiosqlite
aiosmtplib
pyyaml
python-dotenv
httpx
